[#176782168] Api specs refactor
